DefaultPacker takes only uint32 or *uint32 as ID

+++ server_test.go
@@ -164,7 +164,7 @@ func TestServer_handleConn(t *testing.T) {
 		assert.NoError(t, ctx.Bind(&reqData))
 		assert.EqualValues(t, 1, ctx.Message().ID)
 		assert.Equal(t, reqData.Param, "hello test")
-		return ctx.Response(2, &TestResp{Success: true})
+		return ctx.Response(uint32(2), &TestResp{Success: true})
 	})
 	// use middleware
 	server.Use(func(next HandlerFunc) HandlerFunc {
@@ -197,7 +197,7 @@ func TestServer_handleConn(t *testing.T) {
 	reqDataByte, err := codec.Encode(reqData)
 	assert.NoError(t, err)
 	msg := &message.Entry{
-		ID:   1,
+		ID:   uint32(1),
 		Data: reqDataByte,
 	}
 	reqMsg, err := packer.Pack(msg)
@@ -220,7 +220,7 @@ func TestServer_NotFoundHandler(t *testing.T) {
 		Packer: &DefaultPacker{},
 	})
 	server.NotFoundHandler(func(ctx *Context) (*message.Entry, error) {
-		return ctx.Response(101, []byte("handler not found"))
+		return ctx.Response(uint32(101), []byte("handler not found"))
 	})
 	go func() {
 		err := server.Serve(":0")
@@ -237,7 +237,7 @@ func TestServer_NotFoundHandler(t *testing.T) {
 
 	// send msg
 	msg := &message.Entry{
-		ID:   1,
+		ID:   uint32(1),
 		Data: []byte("test"),
 	}
 	reqMsg, err := server.Packer.Pack(msg)
